add to cdk-ffi + move Token to own module
