+++ crates/cdk-ffi/src/lib.rs
@@ -7,6 +7,7 @@
 pub mod database;
 pub mod error;
 pub mod multi_mint_wallet;
+pub mod token;
 pub mod types;
 pub mod wallet;
